premerge

+++ app/Controllers/Http/LocationController.js
@@ -4,28 +4,16 @@ const Location = use('App/Models/Location')
 const Event = use('App/Models/Event')
 class LocationController {
 
-  async getAll () {
-    const location = await Location
-      .query()
-      //.with('partecipants')
-      //.with('category')
-      //.with('location')
-      //.with('creator')
-      .fetch()
-     return location
-  }
+  async getAll ({response}) {
+    const location = await Location.query().fetch()
+      if(location) return response.status(200).send(location)  
+      else return response.status(404).send("Location non esistente")
+    }
 
-  async getId ({request, response, params}) {
+  async getId ({response, params}) {
     try {
       const { id } = params
-      const location = await Location
-      .query()
-      .where('id', id)
-      .with('partecipants')
-      //.with('category')
-      //.with('location')
-      //.with('creator')
-      .fetch()
+      const location = await Location.query().where('id', id).with('partecipants').fetch()
       if(location) return response.status(200).send(location)
       else return response.status(400).send('Location non trovata')
     } catch (error) {
@@ -52,9 +40,11 @@ class LocationController {
       const data = request.all()
       const { id } = params
       const location = await Location.find(id)
-      location.merge({...data})
-      await location.save()
-      return response.send('Location modificata correttamente')
+      if(location){
+        location.merge({...data})
+        await location.save()
+        return response.send('Location modificata correttamente')
+      }else return  response.status(404).send("Location non esistente")
     } catch (error) {
       return response.status(500).send({
         message: error.message
@@ -67,10 +57,12 @@ class LocationController {
     try {
       const { id } = params
       const location = await Location.find(id)
-      await Event.query().where('id_location', params.id).update({ id_location: 0})
-      await location.partecipants().delete()
-      await location.delete()
-      return response.status(200).send('Location eliminata correttamente')
+      if(location){
+        await Event.query().where('id_location', params.id).update({ id_location: 0})
+        await location.partecipants().delete()
+        await location.delete()
+        return response.status(200).send('Location eliminata correttamente')
+      }else return  response.status(404).send("Location non esistente")
     } catch (error) {
       return response.status(500).send({
         message: error.message
@@ -79,18 +71,12 @@ class LocationController {
   }
 
   async getEvents({params, response}){
-    // const location = await Location.find(params.id)
-    const location2 = await Location
-    .events()
-    // .fetch()
-      // .query()
-      // .where('id', params.id)
-      // .with('events')
-      // .with('category')
-      // .with('location')
-      // .with('creator')
-      .fetch()
-     return location2
+    const location = await Location.find(params.id)
+    if(location){
+      const getEvent = await location.events().fetch()
+      return response.status(200).send(getEvent)  
+    }
+    else return  response.status(404).send("Location non esistente")
   }
 }
